Build Warnings Fix
- Fixed C5040 dynamic exception specifications (throw(), throw(std::bad_alloc)) incompatible with C++17 in game and server stdafx.

- Fixed C4838 unsigned int to int narrowing conversion in CharacterClassData::iClassBit (character.h), changed type to unsigned int.

- Fixed C4244 implicit narrowing conversions: double to float in common.cpp, float to int in gamecharacterstatus.cpp, double to int in SkillManager.cpp, double to int/__int64 in server TradeHandler.cpp, wchar_t to char in CWMIReader.cpp and Main.cpp.

- Fixed C4307 signed integral constant overflow in GAME_XOR macro (def.h) by using unsigned literals.

- Fixed C4018 signed/unsigned comparison mismatches in CGameEngine.cpp (time_t vs uint64_t) and CQuestGame.cpp (int vs UINT).

- Fixed C4553 equality operator == used instead of assignment = in CPetSystemHandler.cpp (bug fix: pet system pointer was never being cleared).

- Fixed C4353 constant 0 used as function expression caused by ERROR macro conflict with Windows.h in itemserver.cpp, replaced with LOGERROR.

- Fixed C4101 unused local variable chain in Pikeman.cpp.

- Fixed C4172 returning address of local temporary in UxxxTarget::GetPosition() (HNSSkill.h), changed return type from reference to value.

- Fixed C4700 uninitialized local variable iASMQuestBit in questserver.cpp.

- Fixed C4996 deprecated GetVersionExA calls in CSystem.cpp.

- Fixed LNK4075 /INCREMENTAL incompatible with /OPT:ICF and /LTCG in game project by disabling incremental linking in Release configuration.

- Added GameServer, Login and data to .gitignore (MSBuild intermediate output folders with .obj, .pch, .recipe, .tlog and other build artifacts that contained hardcoded absolute paths).

+++ Server/server/TradeHandler.cpp
@@ -301,10 +301,10 @@ void CTradeHandler::HandleTradeCoin( User * pcUser, User * pcPartner )
 	if ( pcPartner && pcPartner->pcUserData->pcSocketData && pcUser && pcUser->pcUserData->pcSocketData )
 	{
 		//Trade Coins...
-		if ( pcUser->iTradeCoin && COINSHOPHANDLER->HasCoinAmount( pcUser->pcUserData->szAccountName, pcUser->iTradeCoin * 0.90) )
+		if ( pcUser->iTradeCoin && COINSHOPHANDLER->HasCoinAmount( pcUser->pcUserData->szAccountName, static_cast<int>(pcUser->iTradeCoin * 0.90)) )
 		{
 			//Add Coin to Partner...
-			COINSHOPHANDLER->AddCoin( pcPartner->pcUserData->szAccountName, pcUser->iTradeCoin * 0.90 );
+			COINSHOPHANDLER->AddCoin( pcPartner->pcUserData->szAccountName, static_cast<int>(pcUser->iTradeCoin * 0.90) );
 
 			//Discount from user...
 			COINSHOPHANDLER->UseCoin( pcUser->pcUserData->szAccountName, pcUser->iTradeCoin );
@@ -312,12 +312,12 @@ void CTradeHandler::HandleTradeCoin( User * pcUser, User * pcPartner )
 
 			CHATSERVER->SendChatEx(pcPartner, EChatColor::CHATCOLOR_Trade, "> 10 percent of the coins was taxed.");
 
-			CHATSERVER->SendChatEx( pcPartner, EChatColor::CHATCOLOR_Trade, "> You received %s credits from %s!", FormatNumber( pcUser->iTradeCoin * 0.90), pcUser->pcUserData->sCharacterData.szName );
+			CHATSERVER->SendChatEx( pcPartner, EChatColor::CHATCOLOR_Trade, "> You received %s credits from %s!", FormatNumber( static_cast<__int64>(pcUser->iTradeCoin * 0.90)), pcUser->pcUserData->sCharacterData.szName );
 			CHATSERVER->SendChatEx( pcPartner, EChatColor::CHATCOLOR_Trade, "> You now have %s credits!", FormatNumber( COINSHOPHANDLER->GetCoinAmount( pcPartner->pcUserData->szAccountName ) ) );
 
 			CHATSERVER->SendChatEx(pcUser, EChatColor::CHATCOLOR_Trade, "> 10% of the coins was taxed.");
 
-			CHATSERVER->SendChatEx( pcUser, EChatColor::CHATCOLOR_Trade, "> You sent %s credits to %s!", FormatNumber( pcUser->iTradeCoin * 0.90), pcPartner->pcUserData->sCharacterData.szName );
+			CHATSERVER->SendChatEx( pcUser, EChatColor::CHATCOLOR_Trade, "> You sent %s credits to %s!", FormatNumber( static_cast<__int64>(pcUser->iTradeCoin * 0.90)), pcPartner->pcUserData->sCharacterData.szName );
 			CHATSERVER->SendChatEx( pcUser, EChatColor::CHATCOLOR_Trade, "> You now have %s credits!", FormatNumber( COINSHOPHANDLER->GetCoinAmount( pcUser->pcUserData->szAccountName ) ) );
 
 			//Coin Log
